requirements.txt
anthropic
python-dotenv
chromadb
sentence-transformers